Added Talkie Yes/No answer list
Changed Talkie Manager system to use a list of bools for yes/no answers in talkie dialogue, instead of single bool doNotContinueToGame which only tracked one yes or no option per talkie. To use this bool, check TalkieManager's yesNoChoices List, if there was only 1 yes or no prompt answered in the talkie then this list will only have one element, check list[0] to access it. Scales for any number of yes/no options in a talkie. Always do list.Clear() after checking a list to make sure it is empty and accurate for the next talkie